Crate hw_rust

💬 Description

Declare global constants and define the main function.

🚧 Things to Do

Find (or make) a more appropriate Error to return from main.

👀 See Also

🔗 Error

🔗 OnceLock

🔗 Result

📝 Note

I assume these links are correct. Multiple use of names in different namespaces can make the documentation difficult to navigate at times.

Modules

config 🔒

Read the configuration file. It helps if the file exists but, if it doesn’t, a default file should be created at the XDG_CONFIG_FILE location. This is usually ~/.config/<PROGRAM>/config.json.

constants 🔒

files 🔒

getargs 🔒

logging 🔒

utilities 🔒

Define helper functions.

Statics

ARGUMENTS 🔒

CONFIGURATION 🔒

FILE_SYSTEM 🔒

INPUT 🔒

Functions

main 🔒